requirements.txt
amqp==2.1.4
appdirs==1.4.2
billiard==3.5.0.2
celery==4.0.2
dj-database-url==0.4.2
Django==1.10.6
django-appconf==1.0.1
django-ckeditor==5.2.1
django-image-cropping==1.0.4
easy-thumbnails==2.4.1
gunicorn==19.7.1
kombu==4.0.2
livestreamer==1.12.2
Markdown==2.6.8
olefile==0.44
packaging==16.8
Pillow==4.0.0
psycopg2==2.7.1
pyparsing==2.1.10
pytz==2017.2
requests==2.14.2
six==1.10.0
slugify==0.0.1
unicode-slugify==0.1.3
Unidecode==0.4.20
Unipath==1.1
vine==1.1.3
whitenoise==3.3.0
